+++ CHANGELOG.md
@@ -4,7 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [v0.1.6](https://github.com/NoelDeMartin/umai/releases/tag/v0.1.6) - 2024-01-07
+
+### Changed
+
+- Removed Christmas attire.
 
 ### Fixed